Cleanup and add ARM64 arch

+++ src/docker/10-selinuxtools/.prebuild.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 set -e
 
+[ "$FEDORA_BASE_IMAGE" == "" ] && exit 42 # skip image
+# FIXME Remove this when Docker supplies multi-arch base images
+[[ -n "$(docker images -q $FEDORA_BASE_IMAGE 2> /dev/null)" ]] || docker pull $FEDORA_BASE_IMAGE
+docker tag $FEDORA_BASE_IMAGE rancher/os-fedoraconsole-base
+
 cd $(dirname $0)
 
 rm -rf ./build
